+++ src/components/lib/header/index.tsx
@@ -1,27 +1,11 @@
-import { DownOutlined } from '@ant-design/icons';
-import type { MenuProps } from 'antd';
-import { Dropdown, Layout, Row, Space } from 'antd';
+import { Layout, Row, Space } from 'antd';
 import React from 'react';
 import ConectaLogo from '~/assets/conecta-formacao-logo.svg';
 import { store } from '~/core/redux';
 import { setDeslogar } from '~/core/redux/modules/auth/actions';
 import { BoxShadow } from '~/core/styles/colors';
 import ExitButton from '../exit-button';
-
-const items: MenuProps['items'] = [
-  {
-    key: '1',
-    label: <p>Admin 001</p>,
-  },
-  {
-    key: '2',
-    label: <p>Admin 002</p>,
-  },
-  {
-    key: '3',
-    label: <p>Admin 003</p>,
-  },
-];
+import DropdownPerfil from '../dropdown-perfil';
 
 const contentStyle: React.CSSProperties = {
   position: 'sticky',
@@ -41,12 +25,7 @@ const Header: React.FC = () => {
       <img style={{ height: '50px' }} src={ConectaLogo} alt='Conecta Formação LOGO' />
       <Row justify='end' style={{ width: '100%' }}>
         <Space>
-          <Dropdown menu={{ items }} arrow>
-            <Space>
-              Adm COTIC
-              <DownOutlined />
-            </Space>
-          </Dropdown>
+          <DropdownPerfil />
           <ExitButton
             onClick={() => {
               store.dispatch(setDeslogar());
